% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omp.r
\name{get_max_threads}
\alias{get_max_threads}
\title{get_max_threads}
\usage{
get_max_threads()
}
\value{
An integer.
}
\description{
Get the number of threads OpenMP is using.
}
\details{
This is equivalent to calling \code{omp_get_num_threads()} from a parallel
region where the number of threads is unset inside the pragma.
}
\examples{
omp::get_max_threads()

}
\references{
\url{https://www.openmp.org/spec-html/5.1/openmpsu161.html#x205-2380003.7.1}
}
\seealso{
\code{\link{get_num_procs}}
}
