% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omp.r
\name{get_num_procs}
\alias{get_num_procs}
\title{get_num_procs}
\usage{
get_num_procs()
}
\value{
An integer.
}
\description{
Returns the number of processors available.
}
\details{
This number may change from . See the link in the references.
}
\examples{
omp::get_num_procs()

}
\references{
\url{https://www.openmp.org/spec-html/5.1/openmpsu161.html#x205-2380003.7.1}
}
\seealso{
\code{\link{get_max_threads}}
}
