% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omp.r
\name{get_wtick}
\alias{get_wtick}
\title{get_wtick}
\usage{
get_wtick()
}
\value{
A double.
}
\description{
Returns the precision of the timer used by \code{get_wtime()}.
}
\examples{
omp::get_wtick()

}
\references{
\url{https://www.openmp.org/spec-html/5.1/openmpsu186.html#x233-2660003.10.2}
}
\seealso{
\code{\link{get_wtime}}
}
