% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omp.r
\name{get_wtime}
\alias{get_wtime}
\title{get_wtime}
\usage{
get_wtime()
}
\value{
A double.
}
\description{
Returns the the elapsed wall clock time since some arbitrary (but fixed) time
in the past. Therefore, a difference of two of these values gives the wall
clock execution time.
}
\examples{
t1 = omp::get_wtime()
t2 = omp::get_wtime()
t2 - t1

}
\references{
\url{https://www.openmp.org/spec-html/5.1/openmpsu185.html#x232-2650003.10.1}
}
\seealso{
\code{\link{get_wtick}}
}
