% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omp-package.r
\docType{package}
\name{omp-package}
\alias{omp-package}
\title{omp}
\description{
Utilities for setting and reading the amount of compute resources
available to OpenMP. The package wraps all of the OpenMP functions that
would be useful to call from R.
}
\author{
Drew Schmidt
}
\keyword{package}
