% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omp.r
\name{set_num_threads}
\alias{set_num_threads}
\title{set_num_threads}
\usage{
set_num_threads(num_threads)
}
\arguments{
\item{num_threads}{An integer number of threads to use.}
}
\value{
Returns the input \code{num_threads} invisibly.
}
\description{
Set the number of threads for OpenMP to use. Operates via side-effects.
Affects \code{get_num_procs()} but not \code{get_max_threads()}.
}
\examples{
omp::get_num_procs()
omp::set_num_threads(1)
omp::get_max_threads()

}
\references{
\url{https://www.openmp.org/spec-html/5.1/openmpsu120.html#x159-1920003.2.1}
}
\seealso{
\code{\link{get_num_procs}}, \code{\link{get_max_threads}}
}
